Перенесла рендеринг из main.js в презентер + отредактировала main.js, добавила экземпляр класса презентера доски
