+++ src/main.js
@@ -1,18 +1,12 @@
 import SiteMenuView from "./view/site-menu.js";
-import LoadMoreButtonView from "./view/load-more-button.js";
-import BoardView from "./view/board.js";
-import SortView from "./view/sort.js";
-import TaskListView from "./view/task-list.js";
 import FilterView from "./view/filter.js";
-import TaskView from "./view/task.js";
-import NoTasksView from "./view/no-tasks.js";
-import TaskEditView from "./view/task-edit.js";
 import {generateTask} from "./mock/task";
 import {generateFilter} from "./mock/filter.js";
-import {render, RenderPosition, replace, remove} from "./utils/render.js";
+import BoardPresenter from "./presenter/board.js";
+import {render, RenderPosition} from "./utils/render.js";
 
 const TASKS_COUNT = 22;
-const TASK_COUNT_PER_STEP = 8;
+
 
 const tasks = new Array(TASKS_COUNT).fill().map(generateTask);
 const filters = generateFilter(tasks);
@@ -20,86 +14,9 @@ const filters = generateFilter(tasks);
 const siteMainElement = document.querySelector(`.main`);
 const siteHeaderElement = siteMainElement.querySelector(`.main__control`);
 
-const renderTask = (taskListElement, task) => {
-  const taskComponent = new TaskView(task);
-  const taskEditComponent = new TaskEditView(task);
-
-  render(taskListElement, taskComponent, RenderPosition.BEFOREEND);
-
-  const replaceCardToForm = () => {
-    replace(taskEditComponent, taskComponent);
-  };
-
-  const replaceFormToCard = () => {
-    replace(taskComponent, taskEditComponent);
-  };
-
-  const onEscKeyDown = (evt) => {
-    if (evt.key === `Escape` || evt.key === `Esc`) {
-      evt.preventDefault();
-      replaceFormToCard();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  taskComponent.setEditClickHandler(() => {
-    replaceCardToForm();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-  taskEditComponent.setFormSubmitHandler(() => {
-    replaceFormToCard();
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  });
-};
-
-const renderBoard = (boardContainer, boardTasks) => {
-  const boardComponent = new BoardView();
-  const taskListComponent = new TaskListView();
-
-  render(boardContainer, boardComponent, RenderPosition.BEFOREEND);
-  render(boardComponent, taskListComponent, RenderPosition.BEFOREEND);
-
-
-  // По условию заглушка должна показываться,
-  // когда нет задач или все задачи в архиве.
-  // Мы могли бы написать:
-  // tasks.length === 0 || tasks.every((task) => task.isArchive)
-  // Но благодаря тому, что на пустом массиве every вернёт true,
-  // мы можем опустить "tasks.length === 0".
-  // p.s. А метод some на пустом массиве наоборот вернет false
-  if (boardTasks.every((task) => task.isArchive)) {
-    render(boardComponent, new NoTasksView(), RenderPosition.AFTERBEGIN);
-    return;
-  }
-
-  render(boardComponent, new SortView(), RenderPosition.AFTERBEGIN);
-
-  boardTasks
-    .slice(0, Math.min(tasks.length, TASK_COUNT_PER_STEP))
-    .forEach((boardTask) => renderTask(taskListComponent.getElement(), boardTask));
-
-  if (boardTasks.length > TASK_COUNT_PER_STEP) {
-    let renderedTaskCount = TASK_COUNT_PER_STEP;
-
-    const loadMoreButtonComponent = new LoadMoreButtonView();
-    render(boardComponent, loadMoreButtonComponent, RenderPosition.BEFOREEND);
-
-    loadMoreButtonComponent.setClickHandler(() => {
-      boardTasks
-        .slice(renderedTaskCount, renderedTaskCount + TASK_COUNT_PER_STEP)
-        .forEach((boardTask) => renderTask(taskListComponent.getElement(), boardTask));
-
-      renderedTaskCount += TASK_COUNT_PER_STEP;
-
-      if (renderedTaskCount >= boardTasks.length) {
-        remove(loadMoreButtonComponent);
-      }
-    });
-  }
-};
+const boardPresenter = new BoardPresenter(siteMainElement);
 
 render(siteHeaderElement, new SiteMenuView(), RenderPosition.BEFOREEND);
 render(siteMainElement, new FilterView(filters), RenderPosition.BEFOREEND);
 
-renderBoard(siteMainElement, tasks);
+boardPresenter.init(tasks);
